% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{extractCohorts}
\alias{extractCohorts}
\title{Extract exposure cohorts}
\usage{
extractCohorts(
  connectionDetails,
  firstExposureOnly = TRUE,
  exposureTable,
  exposureIds,
  cohortTable = "combined_target",
  resultDatabaseSchema,
  exposureDatabaseSchema = resultDatabaseSchema,
  cdmDatabaseSchema
)
}
\arguments{
\item{connectionDetails}{The connection details to the database. Should be an
object of type \code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{firstExposureOnly}{If a patient appears in more than one exposure
cohort, only keep the earliest exposure, censoring at the start of later
exposure.}

\item{exposureTable}{The table where the exposure cohorts are stored.}

\item{exposureIds}{The cohort definition ids of the exposure cohorts.}

\item{cohortTable}{The table where the combined exposure cohorts will be
stored.}

\item{resultDatabaseSchema}{The database schema where the table with the
combined cohorts will be stored. Needs write permission.}

\item{exposureDatabaseSchema}{The database schema where table with the
exposure cohorts are stored.}

\item{cdmDatabaseSchema}{The database schema where the cdm database is
stored.}
}
\description{
Extract the exposure cohorts from the exposure table. The
function creates a new table where the combined exposure cohorts will be
stored.
}
