% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{generateCohortObservationPeriod}
\alias{generateCohortObservationPeriod}
\title{Generate cohort observation period}
\usage{
generateCohortObservationPeriod(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  resultDatabaseSchema,
  resultTable = "cohort_observation_period",
  maxObservationPeriod,
  washoutPeriod
)
}
\arguments{
\item{connectionDetails}{The connection details to the database. Should be an
object of type \code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{cdmDatabaseSchema}{The database schema where the cdm database is
stored.}

\item{exposureDatabaseSchema}{The database schema where table with the
exposure cohorts are stored.}

\item{exposureTable}{The table where the exposure cohorts are stored.}

\item{resultDatabaseSchema}{The database schema where the table with the new
observation periods will be stored. Needs write permission.}

\item{resultTable}{The table with the new obsrevation periods.}

\item{maxObservationPeriod}{The maximum length of the observation period.}

\item{washoutPeriod}{The minimum number of days before exposure.}
}
\description{
Generates new cohort observation periods for the patients in the
exposure cohorts. At the moment, the new observation period is created by
limiting the actual observation period at the end of the exposure for
each patient.
}
