% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohorts.R
\name{generateModelCohort}
\alias{generateModelCohort}
\title{Generate model cohorts}
\usage{
generateModelCohort(
  connectionDetails,
  cdmDatabaseSchema,
  resultDatabaseSchema,
  resultTable,
  washoutPeriod,
  maxObservationPeriod
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails}.}

\item{cdmDatabaseSchema}{The name of the database schema where the cdm
database is located.}

\item{resultDatabaseSchema}{The name of the schema where the model cohort
table will be stored. Need to have write priviliedges.}

\item{resultTable}{The name of the table that will contain the model cohorts}

\item{washoutPeriod}{The mininum required continuous observation time prior
to index date for a person to be included in the model cohort.}

\item{maxObservationPeriod}{The maximum length of the observation period}
}
\description{
Generate the cohort table for the outcome models
}
