% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateEvents.R
\name{generateNewExposureTable}
\alias{generateNewExposureTable}
\title{Populate new exposure table
Function that creates and populates the simulated exposure table inside the database}
\usage{
generateNewExposureTable(
  connectionDetails,
  exposureDatabaseSchema,
  exposureTable,
  resultDatabaseSchema,
  resultTable,
  eventTimes
)
}
\arguments{
\item{connectionDetails}{The connection details for connecting to the databaseDetails}

\item{exposureDatabaseSchema}{The schema where the exposure table is located}

\item{exposureTable}{The exposure table. Should contain columns
\code{COHORT_DEFINITION_ID}, \code{SUBJECT_ID}, \code{COHORT_START_DATE} and
\code{COHORT_END_DATE}. Column \code{COHORT_DEFINITION_ID} should contain the
exposure cohort definition ids.}

\item{resultDatabaseSchema}{The database schema where the new exposure table
will be stored}

\item{resultTable}{The table with the new exposures}

\item{eventTimes}{A dataframe with the simulated event times}
}
\description{
Populate new exposure table
Function that creates and populates the simulated exposure table inside the database
}
