% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{limitCdmToCohort}
\alias{limitCdmToCohort}
\title{Limit a cdm database to a cohort}
\usage{
limitCdmToCohort(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  outcomeDatabaseSchema,
  outcomeTable,
  resultDatabaseSchema,
  cohortObservationPeriodTable,
  saveDir
)
}
\arguments{
\item{connectionDetails}{The connection details to the database. Should be an
object of type \code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{cdmDatabaseSchema}{The database schema where the cdm database is
stored.}

\item{exposureDatabaseSchema}{The database schema where table with the
exposure cohorts are stored.}

\item{exposureTable}{The table where the exposure cohorts are stored.}

\item{outcomeDatabaseSchema}{The database schema where the table with the
outcome cohorts is stored.}

\item{outcomeTable}{The table with the outcome cohorts.}

\item{resultDatabaseSchema}{The database schema where the table with the
combined cohorts will be stored. Needs write permission.}

\item{cohortObservationPeriodTable}{The table with the new observation
periods.}

\item{saveDir}{The directory where the limited cdm database will be stored.}
}
\description{
Creates a self-contained cdm database based only in the patients
included in a cohort.
}
