% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples.R
\name{runSampleGeneration}
\alias{runSampleGeneration}
\title{Run sample generation}
\usage{
runSampleGeneration(
  simulationDatabaseDir,
  size,
  exposureIds,
  outcomeIds,
  covariateSettings,
  modelDir,
  saveDir
)
}
\arguments{
\item{simulationDatabaseDir}{The directory where a self-contained database
limited to the target patient cohorts is stored.}

\item{size}{The size of the simulated database.}

\item{exposureIds}{The cohort defintion ids of the exposure cohorts of
interest. These are located in the combined cohort table of the limited cdm
database.}

\item{outcomeIds}{The cohort defintion ids of the outcome cohorts of
interest. These are located in the outcome table of the limited cdm
database.}

\item{covariateSettings}{An object of type \code{covariateSettings} generated
with \code{\link[FeatureExtraction]{createCovariateSettings}}.}

\item{modelDir}{The directory where the outcome models are saved.}

\item{saveDir}{The directory where the simulated database will be stored.}
}
\description{
Generates a simulated cdm database.
}
