% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{trainPoissonModels}
\alias{trainPoissonModels}
\title{Train poisson regresion models}
\usage{
trainPoissonModels(
  connectionDetails,
  cdmDatabaseSchema,
  exposureDatabaseSchema,
  exposureTable,
  outcomeDatabaseSchema,
  outcomeTable,
  outcomeIds,
  covariateSettings,
  saveDir,
  ...
)
}
\arguments{
\item{connectionDetails}{The connection details to the database. Should be an
object of type \code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{cdmDatabaseSchema}{The database schema where the cdm database is stored.}

\item{exposureDatabaseSchema}{The database schema where table with the
exposure cohorts are stored.}

\item{exposureTable}{The table where the exposure cohorts are stored.}

\item{outcomeDatabaseSchema}{The database schema where table with the outcome
cohorts are stored.}

\item{outcomeTable}{The table where the outcome cohorts are stored.}

\item{outcomeIds}{The cohort definition ids of the outcome cohorts.}

\item{covariateSettings}{The covariates that will be used for fitting the
poisson regression models. An object of type \code{covariateSettings},
generated with \code{\link[FeatureExtraction]{createCovariateSettings}}.}

\item{saveDir}{The directory where the models will be stored}

\item{...}{Additional parameters passed to
\code{\link[Cyclops]{fitCyclopsModel}}}
}
\description{
Trains poisson regression models
}
